requirements.txt
opencv_python==4.4.0.44
numpy==1.18.5
Flask==1.1.2
Werkzeug==1.0.1
pytesseract==0.3.5
gTTS==2.2.1
Pillow==8.1.2
gunicorn==20.0.4
Jinja2==2.11.2
waitress